Implementation Details

Implementation Details

This section describes important functions and implementation features in greater detail. Additional documentation can also be found in function documentation or in-line.

This section focuses on what the code does and why. Docstrings and the code itself (including comments) provide detailed information regarding how these basic procedures are implemented.

The AbstractModel Type

The AbstractModel type provides a common interface for all model objects, which greatly facilitates the implementation of new model specifications. Any concrete subtype of AbstractModel can be passed to any function defined for AbstractModel, provided that the concrete type has the fields that the function expects to be available.

If a user wants to define a new subclass of models, say regression models, then the user could create a new AbstractRegressionModel type as a subtype of AbstractModel. Functions defined for AbstractRegressionModel would only apply to concrete subtypes of AbstractRegressionModel, but functions defined for AbstractModel will still work on these concrete subtypes.

The AbstractParameter Type

The AbstractParameter type implements our notion of a model parameter: a time-invariant, unobserved value that has significance in the model, e.g. for likelihood computation and estimation.

Though all parameters are time-invariant, they can have different features. Some parameters are scaled for use when solving the model[1] and constructing the model's measurement equations[2]. During optimization, parameters can be transformed from model space to the real line via one of three different transformations. These transformations are also defined as types, and require additional information for each parameter. In some models, steady state values might be relevant parameters. They are typically functions of other parameters, so they do not need to be estimated directly.

These various requirements are nicely addressed using a parameterized type hierarchy.

AbstractParameter{T<:Number}: The common abstract supertype for all parameters.

Parameter{T<:Number, U<:Transform}: The abstract supertype for parameters that are directly estimated.

UnscaledParameter{T<:Number, U:<Transform}: Concrete type for parameters that do not need to be scaled for equilibrium conditions.

ScaledParameter{T<:Number, U:<Transform}: Concrete type for parameters that are scaled for equilibrium conditions.

SteadyStateParameter{T<:Number}: Concrete type for steady-state parameters.

All Parameters have the fields defined in UnscaledParameter:

ModelConstructors.UnscaledParameter — Type.

UnscaledParameter{T<:Number,U<:Transform} <: Parameter{T,U}

Time-invariant model parameter whose value is used as-is in the model's equilibrium conditions.

Fields

key::Symbol: Parameter name. For maximum clarity, key should conform to the guidelines established in the DSGE Style Guide.

value::T: Parameter value. Initialized in model space (guaranteed to be between valuebounds), but can be transformed between model space and the real line via calls to transform_to_real_line and

transform_to_model_space.

valuebounds::Interval{T}: Bounds for the parameter's value in model space.

transform_parameterization::Interval{T}: Parameters used to transform value between model space and the real line.

transform::U: Transformation used to transform value between model space and real line.

prior::NullablePrior: Prior distribution for parameter value.

fixed::Bool: Indicates whether the parameter's value is fixed rather than estimated.

description::String: A short description of the parameter's economic significance.

tex_label::String: String for printing the parameter name to LaTeX.

source

ScaledParameters also have the following fields:

scaledvalue::T: Parameter value scaled for use in eqcond.jl

scaling::Function: Function used to scale parameter value for use in equilibrium conditions.

Note: Though not strictly necessary, defining a scaling with the parameter object allows for much a much cleaner definition of the equilibrium conditions.

Because the values of SteadyStateParameters are directly computed as a function of ScaledParameters and UnscaledParameters, they only require 4 fields:

ModelConstructors.SteadyStateParameter — Type.

SteadyStateParameter{T} <: AbstractParameter{T}

Steady-state model parameter whose value depends upon the value of other (non-steady-state) Parameters. SteadyStateParameters must be constructed and added to an instance of a model object m after all other model Parameters have been defined. Once added to m, SteadyStateParameters are stored in m.steady_state. Their values are calculated and set by steadystate!(m), rather than being estimated directly. SteadyStateParameters do not require transformations from the model space to the real line or scalings for use in equilibrium conditions.

Fields

key::Symbol: Parameter name. Should conform to the guidelines established in the DSGE Style Guide.

value::T: The parameter's steady-state value.

description::String: Short description of the parameter's economic significance.

tex_label::String: String for printing parameter name to LaTeX.

source

The Observable and PseudoObservable Types

We similarly encapsulate information about observables and pseudo-observables (unobserved linear combinations of states, e.g. the output gap) into the Observable and PseudoObservable types. Each type has identifier fields key, name, and longname.

Most importantly, both Observables and PseudoObservables include the information needed for transformations to and from model units. For Observables, these are the input_series, fwd_transform, and rev_transform fields. "Forward transformations" are applied to transform the raw input data series specified in input_series to model units. The model is estimated and forecasted in model units, and then we apply "reverse transformations" to get human-readable units before computing means and bands or plotting. Pseudo-observables are not observed, so they do not have input_series or fwd_transforms, but they may however have rev_transforms.

As an example, the :obs_gdp Observable uses as input_series aggregate nominal GDP in levels, the GDP price index, and population in levels, all from FRED.[3] These series are fwd_transformed to get quarter-over-quarter log growth rates of per-capita real GDP, which are the Observable's model units. The reverse transformation then converts :obs_gdp into annualized quarter-over-quarter percent changes of aggregate real GDP.

ModelConstructors.Observable — Type.

mutable struct Observable

Fields

key::Symbol

input_series::Vector{Symbol}: vector of mnemonics, each in the form :MNEMONIC__SOURCE (e.g. :GDP__FRED). This vector is parsed to determine source (e.g. per-capita consumption gets population and consumption).

fwd_transform::Function: Extracts appropriate input_series from a DataFrame of levels, and transforms data to model units (for example, computes per-capita growth rates from levels).

rev_transform::Function: Transforms a series from model units into observable units. May take kwargs.

name::String: e.g. "Real GDP growth"

longname::String: e.g. "Real GDP growth per capita"

source

ModelConstructors.PseudoObservable — Type.

mutable struct PseudoObservable

Fields

key::Symbol

name::String: e.g. "Flexible Output Growth"

longname::String: e.g. "Output that would prevail in a flexible-price economy"

rev_transform::Function: Transforms a series from model units into observable units. May take kwargs.

source

Model Settings

The Setting type implements computational settings that affect how the code runs without affecting the mathematical definition of the model. Depending on the model, these may include flags (e.g. whether or not to recompute the Hessian), parameterization for the Metropolis-Hastings algorithm (e.g. number of times to draw from the posterior distribution), and the vintage of data being used (Setting is a parametric type - a Setting{T<:Any}, so Booleans, Numbers, and Strings can all be turned into Settings). If settings exist for a model type, then they should be stored centrally in the settings dictionary within the model object.

Why implement a Setting type when we could put their values directly into the source code or dictionary? The most obvious answer is that the parametric type allows us to implement a single interface for all Settings (Booleans, Strings, etc.), so that when we access a particular setting during the estimation and forecast steps, we don't have to think about the setting's type.

Settings play an important role in addition to providing useful abstraction. Estimating and forecasting the New York Fed DSGE model takes many hours of computation time and creates a lot of output files. It is useful to be able to compare model output from two different models whose settings differ slightly (for example, consider two identical models that use different vintages of data as input). A central feature of the Setting type is a mechanism that generates unique, meaningful filenames when code is executed with different settings. Specifically, when a setting takes on a non-default value, a user-defined setting code (along with the setting's value) are appended to all output files generated during execution.

The Setting{T<:Any} type is defined as follows:

ModelConstructors.Setting — Type.

Setting{T}

The Setting type is an interface for computational settings that affect how the code runs without affecting the mathematical definition of the model. It also provides support for non-conflicting file names for output of 2 models that differ only in the values of their computational settings.

Fields

key::Symbol: Name of setting

value::T: Value of setting

print::Bool: Indicates whether to append this setting's code and value to output file names. If true, output file names will include a suffix of the form _code1=val1_code2=val2 etc. where codes are listed in alphabetical order.

code::String: string of <=4 characters to print to output file suffixes when print=true.

description::String: Short description of what the setting is used for.

source

We provide two functions default_settings! and default_test_settings! to initialize settings that most models can have. The settings are

save root

input data root

vintage of data to be used

dataset id

To update the value of an existing function, the user has two options. First, the user may use the <= syntax. However, for this to work properly, it is essential that the setting's key field be exactly the same as that of an existing entry in m.settings. Otherwise, an additional entry will be added to m.settings and the old setting will be the one accessed from other all routines. A potentially safer, though clunkier, option is to use the update! method.

Type Interfaces

Parameter Interface

Base.rand — Method.

Distributions.rand(p::Vector{AbstractParameter{Float64}}, n::Int)

Generate n draws from the priors of each parameter in p.This returns a matrix of size (length(p),n), where each column is a sample.

source

Base.rand — Method.

Distributions.rand(p::Vector{AbstractParameter{Float64}})

Generate a draw from the prior of each parameter in p.

source

ModelConstructors.moments — Method.

moments(θ::Parameter)

If θ's prior is a RootInverseGamma, τ and ν. Otherwise, returns the mean and standard deviation of the prior. If θ is fixed, returns (θ.value, 0.0).

source

ModelConstructors.parameter — Method.

parameter(p::ScaledParameter{T,U}, newvalue::T) where {T<:Number,U<:Transform}

Returns a ScaledParameter with value field equal to newvalue and scaledvalue field equal to p.scaling(newvalue). If p is a fixed parameter, it is returned unchanged.

source

ModelConstructors.parameter — Method.

parameter(p::UnscaledParameter{T,U}, newvalue::T) where {T<:Number,U<:Transform}

Returns an UnscaledParameter with value field equal to newvalue. If p is a fixed parameter, it is returned unchanged.

source

ModelConstructors.parameter — Method.

parameter{T,U<:Transform}(key::Symbol, value::T, valuebounds = (value,value),
                          transform_parameterization = (value,value),
                          transform = Untransformed(), prior = NullablePrior();
                          fixed = true, scaling::Function = identity, description = "",
                          tex_label::String = "")

By default, returns a fixed UnscaledParameter object with key key and value value. If scaling is given, a ScaledParameter object is returned.

source

ModelConstructors.transform_to_model_space — Method.

transform_to_model_space{T<:Number, U<:Transform}(p::Parameter{T,U}, x::T)

Transforms x from the real line to lie between p.valuebounds without updating p.value. The transformations are defined as follows, where (a,b) = p.transform_parameterization and c a scalar (default=1):

Untransformed: x

SquareRoot: (a+b)/2 + (b-a)/2 * c * x/sqrt(1 + c^2 * x^2)

Exponential: a + exp(c*(x-b))

source

ModelConstructors.transform_to_real_line — Method.

transform_to_real_line(p::Parameter{T,U}, x::T = p.value) where {T<:Number, U<:Transform}

Transforms p.value from model space (between p.valuebounds) to the real line, without updating p.value. The transformations are defined as follows, where (a,b) = p.transform_parameterization, c a scalar (default=1), and x = p.value:

Untransformed: x

SquareRoot: (1/c)*cx/sqrt(1 - cx^2), where cx = 2 * (x - (a+b)/2)/(b-a)

Exponential: a + exp(c*(x-b))

source

ModelConstructors.update! — Method.

update!(pvec::ParameterVector{T}, values::Vector{T}) where T

Update all parameters in pvec that are not fixed with values. Length of values must equal length of pvec. Function optimized for speed.

source

ModelConstructors.update — Method.

update(pvec::ParameterVector{T}, values::Vector{T}) where T

Returns a copy of pvec where non-fixed parameter values are updated to values. pvec remains unchanged. Length of values must equal length of pvec.

We define the non-mutating version like this because we need the type stability of map!

source

Setting Interface

Base.:<= — Method.

(<=)(m::AbstractModel, s::Setting)

Syntax for adding a setting to a model/overwriting a setting via m <= Setting(...)

source

ModelConstructors.get_setting — Method.

get_setting(m::AbstractModel, setting::Symbol)

Returns the value of the setting

source

ModelConstructors.update! — Method.

update!(a::Setting, b::Setting)

Update a with the fields of b if:

The key field is updated if a.key == b.key

The print boolean and code string are overwritten if a.print is false and b.print is true, or a.print is true, b.print is false, and b.code is non-empty.

The description field is updated if b.description is nonempty

source

rng::MersenneTwister
testing::Bool
observable_mappings::Dict{Symbol, Observable}
pseudo_observable_mappings::Dict{Symbol, PseudoObservable}

[1]
By solving the model, we mean a mapping from parameters to some objects of interest. In a state space model, solving the model is a mapping from parameters to a state transition function. By constructing

[2]
In the context of a state space model, a measurement equation is mapping from states to observable data.

[3]
In DSGE.jl, we implement a load_data function that parses input_series to retrieve data from FRED. To take full advantage of the Observable type, users may want to write their own load_data function. For example, it may be convenient to write a load_data function that parses input_series to select column(s) from saved CSV files and combines them into a single data frame.